Code moi nhat

+++ EchoMMO-Backend/src/main/java/com/echommo/dto/ExplorationResponse.java
@@ -11,15 +11,16 @@ import java.math.BigDecimal;
 public class ExplorationResponse {
     private String message;
     private String type;       // NOTHING, GOLD, ITEM, COMBAT, REST...
-    private BigDecimal goldGained; // Tổng vàng nhận được (Base + Event)
+    private BigDecimal goldGained;
 
-    private Integer currentExp;
+    // 👇 [QUAN TRỌNG] Đã đổi từ Integer sang Long để khớp với Entity Character
+    private Long currentExp;
+
     private Integer currentLv;
     private Integer currentEnergy;
     private Integer maxEnergy;
     private Integer newLevel;
 
-    // [NEW] Thêm 2 trường này để Frontend hiển thị popup nhận quà
-    private String rewardName;   // Ví dụ: "Túi vàng", "Gỗ", "Kiếm cùn"
-    private Integer rewardAmount; // Ví dụ: 50, 1, 1
+    private String rewardName;
+    private Integer rewardAmount;
 }
